BIG changes reducing ram load

+++ MK1/lib/SignalProcessing/signal_processing.cpp
@@ -1,9 +1,9 @@
 #include "signal_processing.h"
 
 // Variables needed for the Moving Average Filter
-const int numReadings = 10;
-int readings[numReadings];
-int readIndex = 0;
+const uint8_t numReadings = 10;
+uint16_t readings[numReadings];
+uint8_t readIndex = 0;
 float total = 0;
 
 // Variables needed for the Low-Pass Filter
@@ -14,7 +14,7 @@ float alpha = 0.1; // Smoothing factor for low and high-pass filter
 float lastHighPassFiltered = 0;
 float lastInput = 0;
 
-float movingAverage(int newValue) {
+float movingAverage(uint16_t newValue) {
     total = total - readings[readIndex];
     readings[readIndex] = newValue;
     total = total + readings[readIndex];
@@ -23,13 +23,13 @@ float movingAverage(int newValue) {
     return total / numReadings;
 }
 
-float lowPassFilter(int newValue) {
+float lowPassFilter(uint16_t newValue) {
     float filteredValue = alpha * newValue + (1 - alpha) * lastLowPassFiltered;
     lastLowPassFiltered = filteredValue;
     return filteredValue;
 }
 
-float highPassFilter(int newValue) {
+float highPassFilter(uint16_t newValue) {
     // Apply the low-pass filter to the new value
     float lowPass = lowPassFilter(newValue);
     // Calculate the high-pass filter output
@@ -48,7 +48,7 @@ float notchBandwidth = 0.05; // The bandwidth around the notch frequency, as a f
 float r = 1 - 3 * notchBandwidth;
 float k = (1 - 2 * r * cos(2 * M_PI * notchFrequency) + r * r) / (2 - 2 * cos(2 * M_PI * notchFrequency));
 
-float notchFilter(int newValue) {
+float notchFilter(uint16_t newValue) {
     notchOutput = newValue - 2 * cos(2 * M_PI * notchFrequency) * notchOutputPrevious + r * r * notchInputPrevious;
     notchInputPrevious = newValue;
     notchOutputPrevious = notchOutput;
@@ -57,24 +57,24 @@ float notchFilter(int newValue) {
 }
 
 
-const int medianWindowSize = 5; // Window size for median filter
-int medianWindow[medianWindowSize];
-int medianIndex = 0;
+const uint8_t medianWindowSize = 5; // Window size for median filter
+uint16_t medianWindow[medianWindowSize];
+uint8_t medianIndex = 0;
 
 float medianFilter(int newValue) {
     medianWindow[medianIndex] = newValue;
     medianIndex = (medianIndex + 1) % medianWindowSize;
 
     // Create a copy of the window for sorting
-    int sortedWindow[medianWindowSize];
-    for (int i = 0; i < medianWindowSize; i++) {
+    uint16_t sortedWindow[medianWindowSize];
+    for (uint8_t i = 0; i < medianWindowSize; i++) {
         sortedWindow[i] = medianWindow[i];
     }
 
     // Simple insertion sort
-    for (int i = 0; i < medianWindowSize; i++) {
-        for (int j = i; j > 0 && sortedWindow[j - 1] > sortedWindow[j]; j--) {
-            int temp = sortedWindow[j];
+    for (uint8_t i = 0; i < medianWindowSize; i++) {
+        for (uint16_t j = i; j > 0 && sortedWindow[j - 1] > sortedWindow[j]; j--) {
+            uint16_t temp = sortedWindow[j];
             sortedWindow[j] = sortedWindow[j - 1];
             sortedWindow[j - 1] = temp;
         }
@@ -104,7 +104,7 @@ float kalmanFilter(float newValue) {
 }
 
 
-float applyFilter(int bioValue, int filterIndex) {
+float applyFilter(uint16_t bioValue, uint8_t filterIndex) {
   float filteredValue;
   switch (filterIndex) {
     case 0:
